v4l2: link v4l2_buffer and video_buffer video memory upon v4l2_dequeue_buffer
When a struct video_buffer is queued to a video source that has no
backing v4l2 device, it must have some way to access the video memory
that is pointed to by the struct v4l2_buffer. Assign the video memory
pointer of the struct v4l2_buffer to that of the struct video_buffer so
that video sources can access it.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>